test: 로직 수정 및 예상 결제 금액 계산 테스트 코드 작성
BenefitReportTest의 총혜택 금액 계산 테스트를 EventTest로 이동
BeforeEach 메서드로 Order 객체를 생성하도록 로직 변경
