+++ src/test/java/christmas/BenefitReportTest.java
@@ -1,10 +1,5 @@
 package christmas;
 
-import christmas.controller.EventDiscountController;
-import christmas.domain.Order;
-import christmas.domain.OrderedCount;
-import christmas.domain.OrderedMenu;
-import christmas.domain.VisitDate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -13,8 +8,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
 
-import java.util.HashMap;
-import java.util.Map;
 
 public class BenefitReportTest extends NsTest {
     private static final String LINE_SEPARATOR = System.lineSeparator();
@@ -35,21 +28,6 @@ public class BenefitReportTest extends NsTest {
             run("5", "양송이수프-1,레드와인-1");
             assertThat(output()).contains("<증정 메뉴>" + LINE_SEPARATOR + "없음");
         });
-    }
-
-    @Test
-    @DisplayName("총혜택 금액을 계산한다.")
-    void totalBenefitAmount() {
-        VisitDate visitDate = new VisitDate("8");
-        Map<OrderedMenu, OrderedCount> orderResult = new HashMap<>();
-        orderResult.put(new OrderedMenu("티본스테이크", orderResult), new OrderedCount("2"));
-        orderResult.put(new OrderedMenu("레드와인", orderResult), new OrderedCount("1"));
-        Order order = new Order(orderResult);
-
-        int expectedAmount = 30746;
-        EventDiscountController eventDiscountController = new EventDiscountController();
-
-        assertThat(eventDiscountController.calculateTotalBenefitAmount(visitDate, order)).isEqualTo(expectedAmount);
     }
 
     @Override
